dashboard/requirements.txt
streamlit
pandas
numpy
plotly
requests
groq
gtts
SpeechRecognition
pydub
databricks-vectorsearch
